Client/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
google-generativeai
python-dotenv
